requirements.txt
setuptools>=68.0.0
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
pandas>=2.0.0
torch>=2.0.0
torchvision>=0.15.0
ray>=2.6.0
dask>=2023.9.0
distributed>=2023.9.0
tqdm>=4.65.0
multiprocess>=0.70.15
langchain>=0.0.267
langchain-openai>=0.0.2
faiss-cpu>=1.7.4
openai>=1.3.0
tiktoken>=0.5.1
python-dotenv>=1.0.0
chromadb>=0.4.13
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
psutil>=5.9.0
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
peft>=0.4.0
anthropic>=0.2.6
bitsandbytes>=0.40.0
sentencepiece>=0.1.99
tensorboard>=2.13.0
matplotlib>=3.7.0
requests>=2.31.0
scipy>=1.11.0
colorama>=0.4.6
pyyaml>=6.0.0
h5py>=3.8.0

# Additional packages for API service
starlette>=0.27.0

# Additional packages for language models
langchain-community>=0.0.10

# Additional packages for Hugging Face models
huggingface_hub>=0.19.0
aiohttp>=3.8.5
async-timeout>=4.0.3
httpx>=0.24.1

# Notes:
# - pathlib is part of the Python standard library since Python 3.4
# - concurrent.futures is part of the Python standard library
# - fastapi-cors is not needed (CORS is handled by FastAPI's middleware)
# - mpl_toolkits is part of matplotlib
# - zipfile is part of the Python standard library
# - The enum module is part of the Python standard library
# - threading and queue are part of the Python standard library
# - tempfile is part of the Python standard library
# - re is part of the Python standard library

# The posture solver module only uses standard library modules and numpy, which is already included above 